Refactoring env_init as module to be used
while create data type may be called in several places, so let
env_init become a module for other scripts using it to create
data types
